Use EventKind/Variation in the Event derive macro
Move the event parsing into it's own mod to be shared by event.rs.
